moving helpers and texts within overflows source code
