+++ TTS/tts/utils/helpers.py
@@ -236,55 +236,3 @@ def maximum_path_numpy(value, mask, max_neg_val=None):
     path = path * mask.astype(np.float32)
     path = torch.from_numpy(path).to(device=device, dtype=dtype)
     return path
-
-
-def log_clamped(x, eps=1e-04):
-    """
-    Avoids the log(0) problem
-
-    Args:
-        x (torch.tensor): input tensor
-        eps (float, optional): lower bound. Defaults to 1e-04.
-
-    Returns:
-        torch.tensor: :math:`log(x)`
-    """
-    clamped_x = torch.clamp(x, min=eps)
-    return torch.log(clamped_x)
-
-
-def inverse_sigmod(x):
-    r"""
-    Inverse of the sigmoid function
-    """
-    if not torch.is_tensor(x):
-        x = torch.tensor(x)
-    return log_clamped(x / (1.0 - x))
-
-
-def inverse_softplus(x):
-    r"""
-    Inverse of the softplus function
-    """
-    if not torch.is_tensor(x):
-        x = torch.tensor(x)
-    return log_clamped(torch.exp(x) - 1.0)
-
-
-def logsumexp(x, dim):
-    r"""
-    Differentiable LogSumExp: Does not creates nan gradients
-        when all the inputs are -inf yeilds 0 gradients.
-    Args:
-        x : torch.Tensor -  The input tensor
-        dim: int - The dimension on which the log sum exp has to be applied
-    """
-
-    m, _ = x.max(dim=dim)
-    mask = m == -float("inf")
-    s = (x - m.masked_fill_(mask, 0).unsqueeze(dim=dim)).exp().sum(dim=dim)
-    return s.masked_fill_(mask, 1).log() + m.masked_fill_(mask, -float("inf"))
-
-
-
-
